github_jira_tools/requirements.txt
click==6.7
requests==2.20.0
PyGithub==1.43.7
